refactor(logger): Add a hook system with Appender
